Revert "refactor: hooks for Draggable TextBoxfix"
